change: produce values from terms then write those values at appropriate times
rather than running the rules multiple times for each stage, it runs
all of the rules at once, then parcels out what the results of the
rules means. this makes it much more css-like, and makes it easier to
handle a wider variety of terms ( styles ).
previously there was only “before”, “on”, “after” — now, there are:
prefix, content, postfix, quotes, separators, and scope.